\name{weaverOptimiser-package}
\alias{weaverOptimiser-package}
\alias{weaverOptimiser}
\docType{package}
\title{
\packageTitle{weaverOptimiser}
}
\description{
\packageDescription{weaverOptimiser}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{weaverOptimiser}
\packageIndices{weaverOptimiser}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{weaverOptimiser}

Maintainer: \packageMaintainer{weaverOptimiser}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
